move restarting to parent component - #1920
